+++ admin/windows/admin/MainPage.h
@@ -1,0 +1,19 @@
+﻿#pragma once
+#include "MainPage.g.h"
+#include <winrt/Microsoft.ReactNative.h>
+
+namespace winrt::admin::implementation
+{
+    struct MainPage : MainPageT<MainPage>
+    {
+        MainPage();
+    };
+}
+
+namespace winrt::admin::factory_implementation
+{
+    struct MainPage : MainPageT<MainPage, implementation::MainPage>
+    {
+    };
+}
+